test(ivy): remove code duplication from node perf benchmarks (#32104)
PR Close #32104
